pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Comparing conf files."
authors = ["VladislavTolstikov <omfgzwero@gmail.com>"]
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.dev-dependencies]
prompt = "^0.4.1"
flake8 = "^5.0.4"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
